made notifier.next part of std_loop
so that it's not just called in game loop
