+++ pokete.py
@@ -41,7 +41,8 @@ from pokete_classes.notify import notifier
 from pokete_classes.achievements import achievements, AchievementOverview
 from pokete_classes.event import _ev
 from pokete_classes.dex import Dex
-from pokete_general_use_fns import liner, sort_vers, std_loop, parse_args
+from pokete_classes.loops import std_loop
+from pokete_general_use_fns import liner, sort_vers, parse_args
 from release import VERSION, CODENAME, SAVEPATH
 
 
@@ -1250,7 +1251,6 @@ def game(_map):
                 codes(inp)
                 _ev.clear()
         std_loop()
-        notifier.next()
         _map.extra_actions()
         time.sleep(0.05)
         for statement, x, y in zip([figure.x + 6 > movemap.x + movemap.width,
