+++ pokete_classes/buy.py
@@ -2,7 +2,8 @@
 
 import time
 import scrap_engine as se
-from pokete_general_use_fns import std_loop, liner
+from pokete_general_use_fns import liner
+from .loops import std_loop
 from .ui_elements import Box, ChooseBox
 from .inv_items import invitems
 from .event import _ev
