+++ pokete_classes/dex.py
@@ -3,7 +3,8 @@
 import time
 import scrap_engine as se
 import pokete_data as p_data
-from pokete_general_use_fns import liner, easy_exit_loop, std_loop
+from pokete_general_use_fns import liner
+from .loops import std_loop, easy_exit_loop
 from .poke import Poke
 from .event import _ev
 from .ui_elements import ChooseBox, Box
