+++ pokete_classes/movemap.py
@@ -2,7 +2,8 @@
 
 import time
 import scrap_engine as se
-from pokete_general_use_fns import std_loop, liner
+from pokete_general_use_fns import liner
+from .loops import std_loop
 from .classes import OutP
 from .color import Color
 from .event import _ev
