+++ pokete_classes/roadmap.py
@@ -3,7 +3,8 @@
 import time
 import scrap_engine as se
 import pokete_data as p_data
-from pokete_general_use_fns import std_loop, liner, easy_exit_loop
+from pokete_general_use_fns import liner
+from .loops import std_loop, easy_exit_loop
 from .color import Color
 from .ui_elements import Box, InfoBox
 from .event import _ev
