+++ pokete_classes/side_loops.py
@@ -1,6 +1,7 @@
 import os
 import scrap_engine as se
-from pokete_general_use_fns import liner, easy_exit_loop
+from pokete_general_use_fns import liner
+from .loops import easy_exit_loop
 from .ui_elements import InfoBox, StdFrame
