+++ pokete_general_use_fns.py
@@ -1,9 +1,7 @@
 """General use functions for Pokete"""
 
 import sys
-import time
 import release
-from pokete_classes.event import _ev
 
 
 def liner(text, width, pre=""):
@@ -52,22 +50,6 @@ def sort_vers(vers):
         Sorted list"""
     return [k[-1] for k in
             sorted([([int(j) for j in i.split(".")], i) for i in vers])]
-
-
-def easy_exit_loop():
-    """Loops until q or Esc are pressed"""
-    while True:
-        if _ev.get() in ["'q'", "Key.esc"]:
-            _ev.clear()
-            return
-        std_loop()
-        time.sleep(0.05)
-
-
-def std_loop():
-    """Standart action executed in most loops"""
-    if _ev.get() == "exit":
-        raise KeyboardInterrupt
 
 
 def print_help(path):
